visionary/processors/segmentation_processor.py
from typing import Dict, Any, Union
from pathlib import Path
import numpy as np
from PIL import Image
from .base import BaseProcessor

class SegmentationProcessor(BaseProcessor):
    """Processor for segmentation tasks."""
    
    def _load_model(self):  # ← Changed from load_model to _load_model
        """Load segmentation model."""
        print(f"Loading segmentation model: {self.model_config.model_type}")
        self.model = "mock_segmentation_model"
    
    def process(self, input_data: Union[str, Path, np.ndarray, Image.Image], 
                task_config) -> Dict[str, Any]:
        """Process input for segmentation."""
        results = {
            'masks': [[1, 0, 1], [0, 1, 0], [1, 1, 0]],  # Mock mask data
            'labels': ['person', 'background', 'object'],
            'image_size': (640, 480)
        }
        return results
